Escape all required asterisks and underscores
Markdown requires these to be escaped to format them properly if they
are literals. The only functions where they are not escaped are
`inline_code` and `code_block`, as markdwon does not format the content
of code.

It should be noted that manual formating is now more difficult. For
example, to create a header with bold text the following could have been
done before:

    >>> md.header("This be **heavy**", 2)

Now only the following can be done:

    >>> md.header("This be " + md.bold(heavy), 2)

The benefits outweigh the negatives, however.